# project_support/tests/fixtures/original_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SpaceTwin VVP",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/satellites": {
      "get": {
        "summary": "Satellites",
        "operationId": "satellites_api_satellites_get",
        "parameters": [
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 32,
              "default": "active",
              "title": "Group"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50000,
              "minimum": 0,
              "description": "0이면 필터 결과 전체 반환",
              "default": 0,
              "title": "Limit"
            },
            "description": "0이면 필터 결과 전체 반환"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "orbit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(all|LEO|MEO|GEO|HEO)$",
              "default": "all",
              "title": "Orbit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Satellites Api Satellites Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/satellites/{catalog_number}": {
      "get": {
        "summary": "Satellite Profile",
        "operationId": "satellite_profile_api_satellites__catalog_number__get",
        "parameters": [
          {
            "name": "catalog_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 999999999,
              "minimum": 1,
              "description": "NORAD Catalog Number",
              "title": "Catalog Number"
            },
            "description": "NORAD Catalog Number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Satellite Profile Api Satellites  Catalog Number  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}